+++ src/test/java/next/reflection/Junit4TestRunner.java
@@ -2,10 +2,17 @@ package next.reflection;
 
 import org.junit.jupiter.api.Test;
 
-public class Junit4TestRunner {
+import java.lang.reflect.Method;
+
+public class Junit4TestRunner extends AbstractTestRunner {
+
     @Test
     public void run() throws Exception {
-        Class<Junit4Test> clazz = Junit4Test.class;
-        // TODO Junit4Test에서 @MyTest 애노테이션이 있는 메소드 실행
+        super.run(Junit4Test.class);
+    }
+
+    @Override
+    public boolean isTestMethod(Method method) {
+        return method.isAnnotationPresent(MyTest.class);
     }
 }
